# cmd/aznum2words-webapp/api/models/models.gen.go
// Package models provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.12.4 DO NOT EDIT.
package models

import (
	openapi_types "github.com/deepmap/oapi-codegen/pkg/types"
)

// ConvertNumberToWords Convert number to words request representation
type ConvertNumberToWords struct {
	// Number The number to be converted to words
	Number string `json:"number"`
}

// ConvertWordsToNumber Convert words to number request representation
type ConvertWordsToNumber struct {
	// Words The words to be converted to number
	Words string `json:"words"`
}

// Error Representation of an Error that can appear using the application.
type Error struct {
	// Code The code of an error that describes the Error.
	Code string `json:"code"`

	// Message The message of an error that describes the Error.
	Message string `json:"message"`
}

// Number Words representation in number
type Number struct {
	// Number Words representation in number
	Number string `json:"number"`
}

// Word The number representation in words
type Word struct {
	// Words The number representation in words
	Words string `json:"words"`
}

// XCorrelationId defines model for x-correlation-id.
type XCorrelationId = openapi_types.UUID

// ConvertNumberToWordsResponse The number representation in words
type ConvertNumberToWordsResponse = Word

// ConvertWordsToNumberResponse Words representation in number
type ConvertWordsToNumberResponse = Number

// UnknownError Representation of an Error that can appear using the application.
type UnknownError = Error

// ConvertNumberToWordsRequest Convert number to words request representation
type ConvertNumberToWordsRequest = ConvertNumberToWords

// ConvertWordsToNumberRequest Convert words to number request representation
type ConvertWordsToNumberRequest = ConvertWordsToNumber

// ConvertWordsToNumberParams defines parameters for ConvertWordsToNumber.
type ConvertWordsToNumberParams struct {
	// XCorrelationId The unique request identifier
	XCorrelationId *XCorrelationId `json:"x-correlation-id,omitempty"`
}

// ConvertNumberToWordParams defines parameters for ConvertNumberToWord.
type ConvertNumberToWordParams struct {
	// XCorrelationId The unique request identifier
	XCorrelationId *XCorrelationId `json:"x-correlation-id,omitempty"`
}

// ConvertWordsToNumberJSONRequestBody defines body for ConvertWordsToNumber for application/json ContentType.
type ConvertWordsToNumberJSONRequestBody = ConvertWordsToNumber

// ConvertNumberToWordJSONRequestBody defines body for ConvertNumberToWord for application/json ContentType.
type ConvertNumberToWordJSONRequestBody = ConvertNumberToWords